Cambios de desarrollo en pantalla de Turnos. Mejorado el databinding, todavía falta para estar operativa la pantalla.

+++ Controladores/TurnoControlador.cs
@@ -13,7 +13,7 @@ namespace Controlador
             return TurnoData.GetAll();
         }
 
-        public static Turno GetByPrimaryKey(int idTurno)
+        public static Turno GetByPrimaryKey(long idTurno)
         {
             return TurnoData.GetByPrimaryKey(idTurno);
         }
